Fix bad commit in Godot 3.2.1 config
While the commit ea2e976cdd7b8516e881d7ed4588e66601add31 is the exact
one used to build 3.2.1-rc2, it has been displaced in the commit history
and does not appear in `git log` for 3.2...3.2.1. It was included in a
PR by mistake, and that PR was merged after 3.2.1 was released, so it is
now technically a part of the 3.2.2 history. This is a merge commit, so
it's not too too important, and all of its merges are correcty attributed
to 3.2.1 still. But this means the history is forever broken.

The commit in question: https://github.com/godotengine/godot/pull/36342